Bug#27312703 ASAN: HEAP-USE-AFTER-FREE: GREATEST/LEAST FUNCTIONS
The underlying buffer of a String passed to val_str() was reused in the
returned String, although the two Strings themselves were different
objects. Hence, despite being two seemingly independent Strings, freeing
the allocated space in one of them invalidated the buffer of the other.
Because the String comparison function accesses the underlying buffers
of both Strings, both buffers need to be non-overlapping and valid at
time of comparison. Failing to meet this requirement caused
HEAP-USE-AFTER-FREE.

To ensure use of two distinct String objects with non-overlapping and
valid buffers, use a boolean variable instead of pointer comparisons to keep
track of which String object to pass to val_str().

Also changed:
 - Follow the design of other Item classes and reintroduce a String
 member to the Item class for LEAST/GREATEST to avoid copying from a
 temporary buffer after comparison.
 - To improve readability in comparison functions for LEAST/GREATEST,
 use a boolean member variable instead of a comparison sign to determine
 whether arguments should be compared as LEAST or GREATEST.
 - Rewrite signed/unsigned integer comparison in LEAST/GREATEST.

Change-Id: Ica42f259fc2176e9cd00513304b636f9bd7dd6dd