Bug #26680035: CREATE/ALTER USER ERRORS OUT BUT GETS RECORDED IN THE                BINARY LOG
Problem:
--------
Create/Alter user commands having an 'IF [NOT] EXISTS'
clause are written to the binary log even in case of errors.

Analysis:
---------
Before writing to binary log, there was no check whether the
query resulted in error or not. As a result, in case of
errors, the query was still written to binary log.

Fix:
----
The fix checks whether the query resulted in error or not
and in case of errors, the above statements
are not written to binary log.