Bug#27041402 ASSERTION `NEW_VALUE >= 0' FAILED.
Description
-----------
This is a post-push fix to address a '-Wunused-variable' warning when
compiling without the debug option.

Analysis
--------
The problem was introduced because the patch merged two code blocks leading to
the removal of one of them and letting a variable ('one_phase') just being used
for DBUG_ASSERT purposes.

Also, one of the provided test-cases, binlog_xa_gtid_consistency_violation.test,
is failing when ran against a build without the debug option.

Fix
---
- Enclose the given loose variable inside a '#ifndef DBUG_OFF ... #endif' block.
- Add '--source include/have_debug.inc' at the beginning of the test script.