Bug#26990375 WL8117: IF TEMPTABLE ENGINE FILLS DISK, SHOULDN'T CONVERT TO INNODB ON-DISK
When TempTable engine is used to create temporary in-memory
tables and memory limit is reached it allocates the memory
using temporary files and mmap. In case when a temporary file
could not be prepared (e.g. due to lack of disk space) an
error is reported.

For the MEMORY (heap) engine the same error was used to
fallback to a disk temporary table (if there was no memory
left the data was copied to a disk table). As TempTable engine
has its own disk storage mechanism the copy-to-disk-temptable
fallback shall not be used.

The fix disables mentioned fallback for tables created using
TempTable engine and adds a test to check server behavior.

Reviewed-By: Guilhem Bichot <guilhem.bichot@oracle.com>
Reviewed-By: Marcin Babij <marcin.babij@oracle.com>
RB: 18423