Bug#27230463: WINDOW FUNC: CRASH IN DO_COPY_MAYBE_NULL()
Problem:
For a non-nullable field, if it's maybe_null gets set to true because of
is_null_on_empty_table(), we end up having a valid "to_null_ptr" in temp
table and no "from_null_ptr". This results in server exit when we swap the direction
of copying in bring_back_frame_row.

Analysis:
For the bug case, field is marked nullable although it can never be NULL
because aggregation in this case happens in the outer query. But
is_null_on_empty_table() marks the field nullable thinking aggregation
happens within the subquery. As a result, we have a copy_field which has
no from_null_ptr but has a to_null_ptr.
As the from field can never be null, do_copy_maybe_null() tries
to clear the output null bit. Problem arises only in case of windowing functions
as we reverse the copying in bring_back_frame_row().

Solution:
We check for null_bit in do_copy_maybe_null() before clearing the null bit.