Bug#27147095: CONTRIBUTION BY FACEBOOK: ADD RPL_READ_SIZE OPTION
This adds a server variable "rpl_read_size" with a default value of 8kb. It
must be a multiple of 4kb.

This sets the size for reads done from the relay log and binlog. Making it
larger might help with IO stalls while reading these files when they are not
in the OS buffer cache.