Bug#26846549 NODE JOINS GROUP EVEN THOUGH IT IS NOT WHITELISTED ON ONE OF THE GROUP MEMBER
Problem
=======
GCS/XCom implements an all-to-all bidirectional network between the
members of a group.
WL#9345 introduced IP whitelisting support in GCS/XCom, where all
connections coming from a non-whitelisted IP are immediately terminated.
It is possible to create IP whitelists such that, from the point of view
of users, a group is functioning correctly, but in reality, GCS/XCom is
unable to create the all-to-all bidirectional network if some members do
not have all the remaining members whitelisted.
For example, consider three nodes N1, N2, and N3, setup as follows:

* whitelist(N1) = N2, N3
* whitelist(N2) = N1
* whitelist(N3) = N1, N2

Consider that N1 bootstraps the group and N2 and N3 join through N1.
In this setup, N2 does not accept connections from N3.
This leads to a situation where N1 functions as a pivot in the group: N2
only receives N3's application messages via N1.
The group works as intended because it is always possible to create a
majority due to N1: N1-N2 or N1-N3.
If N1 leaves the group, the group becomes unproductive.

Analysis
========
The issue stems from the fact that the IP whitelist is the sole decider
of whether GCS/XCom accepts a connection or not.
An IP whitelist that does not feature all group members prevents the
creation of the all-to-all bidirectional network.

Solution
========
Let GCS/XCom accept incoming connections if the IP is whitelisted *or*
the IP belongs to a current member of XCom configuration.
This way GCS/XCom is always able to create the all-to-all bidirectional
network.
This is done by providing the latest site_def to the
xcom_socket_accept_callback upcall from XCom to GCS.

Reviewed-by: Filipe Campos <filipe.campos@oracle.com>
Reviewed-by: Tiago Jorge <tiago.jorge@oracle.com>
RB: 18395