BUG#20454797 WAIT_FOR_SLAVE_PARAM.INC IS NOT CHECKING FOR ERROR WHILE WAITING THE PARAMETER
Problem and analysis:
The mtr include "wait_for_slave_param.inc" was not detecting errors.
Instead, it would always trigger the timeout waiting for the parameter
if an unexpected error stopped one of the slave threads. This
misbehavior of the include generated timeout errors when other errors
were the real cause of the issue in some test cases.

Fix:
Corrected the code in "wait_for_slave_param.inc" to detect the errors.
Also, added a condition that checks for expected errors, so they can be
safely ignored.
Corrected the test cases that were misusing the mtr include
"stop_slave.inc" to stop slaves when errors were present in either the
connection or the applier thread. These tests now ignore the expected
error.