Bug #27401873 SEGFAULT IN SIMPLECPCCLIENT OF ATRT WHEN RECEIVED COMMAND IS NULL
When ATRT receives a message from CPCD it calls a Parser which tries to find
a command in the message. Regardless of the content of the reply the SimpleCpcClient
would set the output argument user_value to the user_value of the current command.
If the Parser failed to run, the reply would contain a NULL command pointer causing
a SEGFAULT.

Add logic to check the status of Parser before setting the output argument user_value.