Bug#27381901 NDB DATA NODE CRASH ON SENDSIGNAL SIGNAL (GSN: 31, LENGTH: 5, REC BLOCK NO: 0)
There is a race between executing the two signal API_FAILREQ and SCAN_TABREQ at
a TC-block instance. due to the signals traveling different paths to the TC-block.
Thus, even if API_FAILREQ is always sent as the last signal, it may arrive
before a SCAN_TABREQ receive earlier.

---- paste in part of a comment added in the patch ----

It is a requirement that the API_FAILREQ signal is
received as the last signal from the failed 'nodeId'.
It is produced by QMGR and ROUTEed via the TRPMAN(s)
(in the recv-thread) with the intention that it will
enter the job buffers after any signals received
from the failed API node. (Note that the transporter
*is* closed when API_FAILREQ is created, so any more
signals should not arrive from this API node).

However, sending the API_FAILREQ via the proxy
will insert it in another job buffer queue than signals
received from the API-node:

  API-request :       recv/TRPMAN-thread  ---> Block-instance
                                              ^
  Routed-API_FAILREQ: recv/TRPMAN-thread     /
                                     \      /
                                      Proxy

Depening on the order the queues are processed, the
API_FAILREQ may then be processed by the Block-instance
before a API-request from the failed node - Which is
not what we want.

Thus we let any proxies receiving a API_FAILREQ, route it
back to the recv/TRPMAN. There it will be inserted in
the same queue as the API-requests, and thus remove the
posibilities for being overtaken:

  Routed-API_FAILREQ: recv/TRPMAN-thread
                                    \
 re-route to TRPMAN:  -----------<-Proxy (we are here)
                      |
                      v
 API-request+FAIL :   recv/TRPMAN-thread  ---> Block-instance
------------- end ---------------

As the comment says, this patch fixes the race by letting any proxy-block
seeing a API_FAILREQ signal route it back via the TRPMAN, where it will enter
the same job-queue as the SCAN_TABREQ. Thus the sequence is guaranteed from
there on.