Bug#27308632 NDB NODE RESTART SHOULD FAIL IF FILESYSTEM NOT UPGRADED
Version 7.6.4 introduces some changes to the disk data tablespace file
format, and the Local Checkpoint LCP disk file format. Due to this any
upgrade of the data nodes to v7.6.4 should use a initial restart to
create the necessary files in the NDB file system. If initial restart
is not done, the data nodes will crash during upgrade failing to find
the required files. To avoid this crash, this patch adds a check and
fails with a clear error if an initial restart is not done. This also
prevents any damage being done to the data on disk. This commit has the
fix and the upgrade test case is pushed in a lower version and merged
up.

(cherry picked from commit ac7dd7ed27bd644ff6009cec9ff08333db7d7d1c)