Bug#27418633 NDB.NDBINFO DEPEND ON NDB TABLE ID
In MTR the Ndb table id is not stable between runs of a test case.

Table id may depend on earlier test runs.

Mostly that is caused by creation of ndb_index_stat tables are delayed.

Uses of table id in ORDER BY clauses are replaced.

In some cases the new order will be different and test result reordered.

Ordered comparisions with table id in WHERE clauses are also replaced.

In some queries the name of dependent tables including the base table id
in name had that id replaced.

But to ensure that a only complete table id in names are replaced
condition on context of replacement was added.

Instead of replace any substring matching <table_id>, for blob tables
_<table_id>_ was replaced and for index tables /<table_id>/.

In some queries to not be dependent of the existence of ndb_index_stat
tables all tables in mysql schema are excluded.