Bug#27245598: MYSQLTEST: EXECUTION PHASE WARNINGS OVERWRITE THE PREPARE               PHASE WARNINGS
Issue:
------
To avoid printing duplicate warnings, there is code in mysqltest which
sets the prepare phase warnings to NULL if execute phase warnings are
present. However, there might be different warnings generated during
prepare and execute phases. In this case, the prepare phase warnings
will be missing from the result file.

Fix:
----
If there are prepare phase warnings, as well as execute phase warnings,
check whether the prepare phase warnings are the same as the execute phase
warnings. If a prepare phase warning is not present in the execute phase,
then print it.

Reviewed-by: Pavan Naik <pavan.naik@oracle.com>
Reviewed-by: Srikanth BR <srikanth.b.r@oracle.com>