Bug #27233287: WINDOW FUNC: CRASH IN ITEM::SAVE_IN_FIELD_INNER
Problem
"maybe_null" of the window function is not set to true when the argument's
"maybe_null" gets set to true once it is determined that this argument is
part of group by and as a result it can have a NULL because of a ROLLUP
operation.

Solution:
Set "maybe_null" of windowing function's argument to true if it
is part of group by. This is in similar lines to how we do
for non-aggregate-functons.