Bug #25822988 ENABLE DIRECT MONITORING OF BACKUP BUFFER LOG  USAGE
Fix for Windows build failure. The build failed because the string
'BACKUP_DATA' is specified as an enum value, but this string is
already in use. The build is fixed by renaming the enum value to
BACKUP_DATA_BUFFER. The  other new enum value of BACKUP_LOG is
also renamed to BACKUP_LOG_BUFFER for consistency.