Bug#25968185: INIT_CONNECT PREVENTS CONNECTIONS IF PASSWORD               EXPIRED
Description: Mysql server restricts client connections from
users with expired passwords when the server variable,
"--init_connect", is set.

Analysis: Mysql server will try to execute the SQL statements
set to the server variable "--init_connect" for each client
that try to connect to the server. When users with expired
password try to connect, mysql server fails to execute the
SQL statements set to "init_connect" which in turn restricts
the user to connect to the server.

Fix: As part of the fix, it has been decided to allow users
with expired password (similar approach is taken for users
with SUPER privilege) to connect to the server without
executing the "init_connect" statements. A warning  is
printed to the server log that "init_connect" statements
were skipped due to the expired password.