Bug#27374791: RESET PERSIST DOES IMPLICIT COMMIT AND DOESN'T NEED RELOAD               PRIVILEGES
Problem: RESET PERSIST is not an administrative command. However since
RESET sql statement is considered as an administrative command there is
an implicit commit when RESET PERSIST is executed.

Fix: stmt_causes_implicit_commit() this function returns TRUE for all
statements which needs to do an implicit commit. In this function we
check if it is RESET command and if it is RESET PERSIST then return
FALSE.