Bug#27087547: ASSERTION FAILED: MAP->BITMAP && MAP2->BITMAP WHEN BUILD -DMAX_INDEXES=255
TABLE_SHARE objects are not constructed by calling a
constructor, which means that all members must be
initialized by the code allocating the memory. The bitmap
TABLE_SHARE::visible_indexes was not initialized in some
cases. Fixed by initializing it whenver the other bitmaps
are initialized.

There is no test case, since you have to build with
MAX_INDEXES=255 in order to trigger the bug.

Change-Id: I205d14a42af6b68d8e4d0138a1e772a550051f13