Fix for Bug#26236857 LRU_GET: ASSERTION `!IS_BUSY_MACHINE(&RETVAL->PAX)' FAILED
The assertion fails begause lru_get returns a Paxos instance that is busy. The sequence of events that lead to this is the following:
Node 1 sends all of the messges which have been decided to node 2, which is doing recovery. This causes all those instances to be moved to the head of the LRU, leaving the busy instances as the next victims for replacement.

This patch fixes lru_get so it skips any busy instances. In addition, the log sending code has been changed so it will not rearrange the LRU.

Doing this solves the reported problem, but leaves a residual problem, which can be reroduced by running the repro test case described in the bug report. Node 2 will refuse to join, with  the message "ERROR 3092 (HY000) at line 3: The server is not configured properly to be an active member of the group. Please see more details on error log."

If we run the test case as desribed in the bug report, with load, the reason it fails is that node 1 sends a snapshot from that point where node 2 originally joined. Since node 1 do not have all the log messages, but only the tail of the log, this leaves a gap. This causes node 2 to time out, since xcom cannot deliver all the messages it needs.

It we run the test with no load, it starts at the same point, but this time node 1 has all the messages. Node 2 now fails with [ERROR] Plugin group_replication reported: 'This member has more executed transactions than those present in the group. Local transactions: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa:1-5 > Group transactions: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa:1-4' This is happens because GR on node 2 now gets the original messages a second time.

The problem can be traced back to the changes done in "BUG#23718481 - A NODE THAT LEAVES AND JOINS WITHIN A 5 SECOND WINDOW MAY NOT GET A GLOBAL VIEW"
This change sets the starting point in the log to the last modified config LSN, which works fine as log as the config has been modified before node 2 rejoins. This is not the case here, since node 1 does not have majority alone, and thus is unable to evict node 2 from the group.

So, the third and last change in this fix is to introduce an alternative solution to the problem described in BUG#23718481 which does not depend on the config modification. We send an internal unique Xcom instance ID as part of the pax_msg, and use this to detect that a new instance of xcom has appeared. The detector_task will send a new global views when a new instance has been detected. This ensures that the new node will get a global view.