Bug #25902137 NDBAPI CRASH IN CHECKMAGICNUMBER DURING OBJECTMAP LOOKUP
An objectmap entry can either be an NdbTransaction pointer or
an NdbReceiver pointer. The ndbapi client receives an invalid
signal which contains a reference to an NdbReceiver objectmap
entry. The referenced entry has already been deleted and a new
NdbTransaction entry has been inserted at the same location.
The objectmap lookup returns an NdbTransaction pointer, which
can then be referenced in 2 ways, both of which cause a crash:

- dereference the pointer as an NdbReceiver to get the owner
of the NdbReceiver

- assert on the magic number check which determines that the
object is a valid NdbReceiver

Fix: After an objectmap lookup, always check the magic number
before proceeding. An invalid magic number triggers handling
for invalid signals, not a crash.