Bug#27397802 ENTIRE CLUSTER CRASHES WITH "MESSAGE TO BIG IN SENDSIGNAL" WITH BIG IN CLAUSE
A large In-list caused a pushed condition exceeding the max 32K limit
for an un-fragmented sendSignal to be created. This caused all datanodes
to crash when the SPJ block sent this condition as part of a LQH_KEYREQ
signal.

The advantage of having such a pushed condition as part of a 'REQ'
returning at most a single row is questionable on its own. As in
this case the size of the LQH_KEYREQ with the pushed condition, can
in some cases be larger than the saved size of the eliminated row.

Furthermore, the condition was also reevaluated on the mysqld server
when arriving there, so nothing saved on the server side either
by pushing the condition to the datanodes.

This patch simply omit sending of the pushed condition as part of
a *lookup operation* (EQ_REF) being part of a pushed join.

Be aware that explain of the query still shows the condition as
being pushed. That is another existing Bug:

 27429615 QUERY EXPLAIN SHOWS PUSHED CONDITIONS WHICH ARE NOT PUSHED.