BUG#27436052	MEMORY LEAKS IN GCS UNIT TESTS
Problem
---------------------------------
The following unit tests are failing with memory leaks in ASAN:
gcs_parameters
gcs_whitelist
gcs_xcom_interface_unit

Compile with -DWITH_ASAN=1, run the tests directly from the runtime output
directory or just run 'ctest -R "gcs*"'

Analysis
---------------------------------
This is caused by successive calls to Whitelist configure() method, now that the
Whitelist has objects instead of std::vector. It was clearing the list instead
of doing a list cleanup.

This is not observable in Group Replication since the object is always created
and destroyed in subsequent calls.

Fix
---------------------------------
Add a clear() method to the whitelist and call it in configure(). Refactor the
destructor to also use the new method.