Bug#27294009 - XCOM ENTERS AN INFINITE LOOP IN CASE OF A REQUEST_RETRY
Description:
-----------
When a node tries to enter a group, its request might be rejected by the
node that is receiving the request. This may happen for a multitude of
reasons and the reject message returns two possible values:
- REQUEST_FAIL: this represents a full rejection.
- REQUEST_RETRY: this represents a "please try again in 1 second". This
exists for the case in which a node that wishes to join the group contacts
a seed node that is itself still booting within a group.

This problem can manifest itself when, for instance, one misconfigured the
group by not adding the bootstrap flag = true. In that case, that node can
enter an infinite retry loop, since currently a node will continue to retry
until it gets a REQUEST_FAIL or a REQUEST_OK. GR will detect this issue and
call leave() upon GCS, but that request will fail because XCom is stuck
in the infinite retry loop.

Solution
--------

This patch adds a limit to the number of retries that a node is allowed
to execute. After the limit is reached, the node will stop retrying and
will return with an error.