Bug #26625652: SIG=11 ROW_SEL_SEC_REC_IS_FOR_CLUST_REC ROW_SEARCH_MVCC
Problem:
In tables with compact row format blob-like types (e.g. JSON) are
stored on separate pages. If another connection is reading data
in read uncommitted isolation mode it may happen that the lob data
is not yet stored. In that case it is not possible to generate
virtual column valu that depends on such lob (e.g. for column
defined using json_extract function). One of paths was not handling
such case correctly, leading to a crash.

Solution:
In the broken path the virtual column generation error is handled
and the error is propagated to be returned by the executed statement.
Such approach was selected to be consistent with other paths already
returning errors in that cases.

RB: 18574
Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>