Bug#22279903 ISSUE WITH ONLY_FULL_GROUP_BY AND UNIQUE KEY WITH NULL VALUES
This is a user-requested improvement to the detection of functional
dependencies in queries with GROUP BY (introduced in 5.7 by WL#2489).
Considering:
CREATE TABLE t(a INT NULL, b INT NOT NULL, c INT, UNIQUE(a,b));
such index doesn't enforce uniqueness on NULLs, so allows, for example,
(NULL,1,1) and (NULL,1,2). So it is incorrect to accept
SELECT a,b,c FROM t GROUP BY a,b;
due to possible duplicate NULL values of 'a' ;
but this query is deterministic:
SELECT a,b,c FROM t WHERE a IS NOT NULL GROUP BY a,b;
and can be accepted (the standard doesn't request that, though).
The goal of this patch is to accept it and variants.
We implement that if a join condition or WHERE clause
is known to never be true when a column 'a' is NULL, then this column can
be considered "known-not-nullable" and thus, if it's part of an index,
we may consider the index as a primary key, which may determine
new functional dependencies.
That's the case in the last query: "a" is known not nullable, so we
may consider "(a,b)" as PK so "c" is determined.
Example conditions: =, <=, >, between, is not null.
This may deserve a paragraph in the manual.

Implementation:
- in analyze_conjunct() we already went through the AND-ed conditions,
but considered only equalities
- now we consider other predicates (Item_func), and if we discover that
we can treat column 'a' as non-nullable, and 'a' is already functionally
dependent, then we mark the Item (with a new value in the 'marker'
member)
- when we look at indexes, we do not exclude indexes with nullable columns
anymore, if the nullable columns are marked as known-not-nullable.

Long comments to explain when exactly we can treat a column as non-nullable
(devil is in the details!) are added in aggregate_check.h.

As Item::marker gets a new usage, cleaned the member up:
made it robust against uses by multiple successive phases, and documented
the rules (see item.h). Changed it from int32 to enum for easier
grep-ping.