Bug #26003816 	RPL.RPL_IPV4_AS_IPV6 AND RPL.RPL_IP_MIX FAILS SPORADICALLY ON PB2
Problem
=======
When the mysql_socket_bind(...) call on the INET socket fails, mysqld
reports an error like that:
[ERROR] Can't start server: Bind on TCP/IP port: No such file or directory
[ERROR] Do you already have another mysqld server running on port: 3310 ?
[ERROR] Aborting

Analyse
=======
Sometimes the port is not released fast enough when stopping and
restarting the server. This happens quite often with the test
suite on busy systems. Retry to bind the address at some intervals
during the port timeout. The value of port timeout is 0 in the
test. So the above error happens if the first try of binding the
address fails.

Fix
===
To fix the problem, set --port-open-timeout to 10000 to allow
enough retries to bind the address.