Bug#27450520 NDB : TR::PERFORMSEND() BROKEN
TransporterRegistry::performSend() is assuming that the
theTransporters array is densely filled in transporter
creation order, to minimise the number of transporters
which must be examined.

However it is not, it is indexed by remote node id.

This means that performSend() will not attempt to send
to high numbered node ids if the node ids are sparsely
allocated.

To fix this, a new dense array of Transporters is added.

performSend() is modified to use the new array.