Bug #27461752 NDB-API, INDEXSCANOPERATION, INCORRECT RESULT WHEN NULL-BOUND IS SPECIFIED
Code generated for NdbIndexScanOperation::setBoundHelperOldApi() failed to set
the 'nullBit' mask correctly if a NULL value was specified for the 'bound'.

Was caused by 'delete-null-pointer-checks' optimization in gcc compiler,
in combination with memcpy() having the 'nonnull' property set for its two
pointer arguments. Thus it assumed that 'aValue != NULL' and eliminated
later code checking for 'aValue == NULL'