BUG#27319084 - INDEX CORRUPT: INNODB: FIELD 0 LEN IS NNN, SHOULD BE NNN; RECORD
This corruption happens under these conditions:
1. Server now has utf8mb4 as default character set, so storage of CHAR(n)
would be n * 4
2. The table has one or more virtual columns defined as CHAR
3. After an ALTER TABLE which changes the table row format, from compact
to non-compact(some problems exist from non-compact to compact too).
4. During the ALTER TABLE, there is concurrent INSERT or UPDATE to the
old table.

After 1, the data type for the virtual columns mentioned in 2 would become
DATA_MYSQL, instead of DATA_CHAR. So for DYNAMIC/COMPACT row format,
InnoDB will try to remove trailing spaces before storing, while for
REDUNDANT format, it doesn't. So let's say the ALTER TABLE changes the
row format from DYNAMIC to REDUNDANT, a concurrent INSERT will log online
operation in DYNAMIC format, but to apply the INSERT to the new table,
the DYNAMIC format for virtual columns of CHAR isn't consistent with the
new definition, due to missing trailing spaces.

The fix should adjust the virtual column values from the online operation
logs. If the new table is of non-compact(REDUNDANT) format, and the logged
value length is less than the column definition, it has to add trailing
spaces to make it a true REDUNDANT format value. If the new table is of
compact(DYNAMIC/COMPACT) format, and the logged value has trailing spaces
due to original REDUNDANT format, it has to remove all the trailing spaces.
The latter is about both correctness and disk space saving.

RB: 18613
Reviewed by: Jimmy Yang <jimmy.yang@oracle.com>