Bug#88825    Contribution: Export zlib version in a variable
If [innodb_log_compressed_pages] is used one needs to be careful which
zlib version in used when using InnoDB recovery. Among others this
could impact backup/restore with a hot backup tool or a storage
snapshot.

This patch exposes the zlib version.

The patch is based on a contribution from Daniel van Eeden (Github user: dveeden)

Change-Id: Idd2481696c40c8feb61720143243ac953bd57f1b