Bug#27265584: minimized serialized dictionary information
Problem: The rapidjson::PrettyWriter produces nicely formatted SDIs
which are suitable for human readers. But the resulting SDIs are almost
twice as large as when using the ordinary writer which skips
unnecessary whitespace. For shipping between nodes in NDB and storage
in tablespaces a smaller size would be prefereable.

Change-Id: Ic34f01505f06d904f260744324556ad9a7ca7121
Solution: Change the typedef for the Sdi_writer so that the ordinary
rapidjson::Writer is used.