BUG#27109556: ADD SERVICE NOTIFICATION VIA SOCKET
Problem: When running mysqld as a SYSV daemon, systemd is not always
able to infer the correct state of the double-forked, real, mysqld
process.

Solution: Add code to allow mysqld to write its state to the systemd
notify socket, in the format expected by systemd. This will allow
systemd to run mysqld as a normal process (Type=notify).

If the name of the socket file is specified in the environment variable
NOTIFY_SOCKET mysqld will attempt to open connection and write its state
changes there. If connecting or writing fails for any reason, a warning
is printed in the error log, but no further action is taken.

Change-Id: Ibb962d22232e86315dbfec6b5ce160153638f56d