Bug#25894101: TEST CASE I_INNODB.INNODB_BUG16244691 FAILED IN SHUTTING DOWN THE SERVER.
Problem: Innodb purge threads could end up in a situation where they
needed to process a large number of log records referring to dropped
tables, which do not exist in the DD. The purge thread looks up the
tables in the DD using the se_private_id, and this would take a long
time since the DD would need to access both the TABLES and PARTITIONS
tables on disk to verify that the table did not exist. This could lead to
timeouts, e.g. during shutdown.

Solution: Add a cache of se_private_ids known not to be associated
with a table. The cache must be an lru of fixed size, since there is
no simple way to determine when an se_private_id will no longer be
used.

Change-Id: I9510da9116f26bbbb3aa00a0abdba0e63ea27503