Follow-up to fix for bug #25915132 "INPLACE ALTER TABLE WITH FOREIGN KEYS CAUSES TABLE DEFINITION MISMATCH".
Fix for nist.nist_all test failures due to assertions.

The problem was that we didn't lowercase generated foreign
key name when acquiring MDL on them (this is necessary as
FK names are case-insensitive and MDL subsystem treats all
keys as binary). As result we got assertion failures when
we tried to create FKs for table with name in upper case or
rename table with FKs to such name.

This patch fixes the problem by ensuring that we use lowercased
version of generated FK names for acquiring MDL.

Additional test coverage for such cases was added to
main.foreign_key test.