Bug#26171471: SIG11 IN LOG_SERVICE_INSTANCE_RELEASE_ALL() IN LOG_BUILTINS.CC
Change loadable logging components to use REQUIRES_SERVICE...
boilerplate so they may benefit from improved locking in
component registry.

With the previous locking, stale entries for services could
remain in the registry when used with a very high number of
concurrent threads, and the logging framework would then try
to utilize stale data, and fail to log as the alleged components
had in fact been removed.

With the registry changes in place, we can no longer easily use
acquire() / release() from a service (but may do so from a plug-in
that only uses, but does not provide services), as our de-init
function is now called while the registry lock is held, and our
release would try to acquire that lock again. Therefore, components
are updated to use the more recent REQUIRES_SERVICES boilerplate.

Patch also updates log_broadcast() (setting of log-related
system-variables) so it uses the cached log_services rather
than acquire them dynamically. This once again removes a
dependence on acquire()/release() (see above).

Further, we split get_server_errmsg() into error_message_for_error_log()
and error_message_for_client() to make the intent of the API clearer
and bring it in line with the ongoing efforts to separate error messages
intended for the error-log and those intended for the client as well
as their respective locale settings.

Finally, the patch preserves the log-line's default iterator
during log-event submission to improve robustness when the
logger calls a log-service, and the log-service in turn calls
the logger again (which is legal behaviour).