Bug#27452115: ASSERTION FAILED: (UCHAR*) TABLE->DEF_READ_SET.BITMAP                 + TABLE->S->COLUMN_BITMAP_
In an error case, bitmaps of read and write set were not reset correctly
resulting in an assert.

Error in this case is not generated in copy_fields(). Its in ::val_int()
called from after_frame(). But we do not catch the error there. Instead, its
caught in copy_fields where the error check happens.

We reset the bitmaps correctly in bring_back_frame_row() for an error case too.