Bug#24488729: CHANGE TYPE OF COPY_FUNCS
Refactoring inspired by bug no 23196195. The List
Temp_table_param::copy_funcs was declared to contain Item's,
but was only ever populated with Item_copy's, and used a
C-cast to cast them back down. This is hereby changed to use
std::vector<Item_copy>, copy_funcs is renamed to
grouped_expressions and it is documented. We also change
Temp_table_param::copy_fields to be an std::vector.

Change-Id: I71072c58e60cf2b023b125edcd0c861ea6032996