Bug#27492656: MTR IGNORES THE INIT_FILE OPTION WITH --BOOTSTRAP DURING SERVER INITIALIZATION
Issue:
------
If the --init-file option is given as a value to the --bootstrap option in MTR,
the SQL statements in the file must be executed during the server initialization
time. However, MTR has a bootstrap.sql of its own which it uses during the
initialization process, and that overwrites any sql file which is passed on the
command line by the user.

Fix:
----
If --init-file option is passed during the initialization process, then the
contents of this sql file are appended to the end of the bootstrap.sql file
which MTR uses.

Change-Id: Iff62eb0859c56806dae086fdfb79bd8b3e154b78