BUG#27317431: LAST_CONFLICT_FREE_TRANSACTION SHOWS DIFFERENT VALUES IN GROUP (EMPTY ON JOINER)
Upon group replication, the LAST_CONFLICT_FREE_TRANSACTION of the
performance_schema.replication_group_member_stats table is empty
(because that table is reset). This empty value is not broadcasted or
accepted by the other members of the group, resulting in a wrong view
when querying this table on the other members of the group (until
a transaction is executed again).

The underlying issue was two-fold. First, the empty transaction
identifiers weren't being encoded in the stats message nor being
accepted in the stats message reception callback. Second, the empty
value was being used to signal  whether or not to send the transaction
identifiers. Since empty values are acceptable values, they can't be
used for this purpose. Given that the stats message is sent every second
any way, we maintain an old copy of the transaction identifiers and only
update them every 30 seconds (which is the original period for this
update).