Bug#27499518 CHANGE THE HIDDEN ATTRIBUTE FOR COLUMNS TO AN ENUM
In order to implement functional indexes, we need to mark a column as hidden
from users. We already have a field named "hidden" for columns, but it hides
the column completely from the optimizer. We need have a column that is
visible to the optimizer, but invisible to the user.

This bugfix change the field "hidden" from a BOOL to
ENUM('Visible', 'SE', 'SQL'), so we can have different levels of visibility.
This will also make it more consistent with the field "hidden" for tables,
which already is an ENUM.

Change-Id: Id4289ee6b4a88cb94dc4099349d1f09a2002f5a0