BUG#20677683 TRANSACTION BOUNDARY PARSER ASSERTS FOR UNKNOWN EVENT TYPES              WHEN AUTO_POSITION=1
The transaction boundary parser is failing a slave assertion when
replicating unsupported non-ignorable events from the master.

This assertion only exists in debug builds. In non-debug binaries,
the transaction boundary parser keep doing its job trying to guess the
end of the current transaction by relying on usual transaction
termination events.

The assertion was designed to ensure that in supported replication
environments, errors in the transaction boundary parser are only
expected if the replication channel is not using MASTER_AUTO_POSITION,
but we are now removing the assert and making the debug server behave
the same way non-debug does.